correctly normalize json path
in the case where you make a relative require w/ a `..` in the path:

```js
require([
    'json!./../string-bundle.json'
], function(strings) {
    // ...
});
```

the plugin's `normalize` method needs to call the supplied `normalize`
parameter (i.e. [the second
argument](http://requirejs.org/docs/plugins.html#apinormalize) to the
plugin's `normalize` property) in order to correctly resolve the `..`.